requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
pyjwt
bcrypt==3.2.0
billiard==3.6.4.0
blinker==1.4
celery==5.2.7
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.8
flask
Flask
Flask-SQLAlchemy
Flask-WTF==1.0.1
future==0.18.2
greenlet==2.0.1
Jinja2==3.1.2
Mako==1.1.3
marshmallow==3.19.0
redis==4.4.0
requests==2.25.1
SecretStorage==3.3.1
six==1.16.0
SQLAlchemy==1.4.45
sqlparse==0.4.3
WTForms==3.0.1
xdg==5
zipp==1.0.0
